bot-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
langchain>=0.0.325
langchain-community>=0.0.1
langchain-openai>=0.0.1
openai>=1.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0 